requirements.txt
fastapi>=0.104.1,<0.115.0
uvicorn[standard]>=0.24.0,<0.32.0
python-multipart>=0.0.6,<0.1.0
pydantic>=2.5.0,<3.0.0
whisperx==3.1.1
torch>=2.4.0,<2.8.0
torchaudio>=2.4.0,<2.8.0
moviepy>=1.0.3,<2.0.0
python-ffmpeg>=2.0.12,<3.0.0
requests>=2.31.0,<3.0.0
aiofiles>=23.2.0,<25.0.0
python-dotenv>=1.0.0,<2.0.0
Pillow>=10.1.0,<12.0.0
numpy>=1.24.3,<2.0.0
